refactor(server): flatten wallet domain layout
