+++ apps/server/src/domain/wallets/index.ts
@@ -1,9 +1,21 @@
 export * from "./domain-errors";
 export * from "./models";
+export * from "./repository/payment-attempt.repository";
 export * from "./repository/wallet-hold.repository";
 export * from "./repository/wallet.repository";
-export * from "./services/stripe-webhook.service";
-export * from "./services/wallet-hold.service";
-export * from "./services/wallet.service";
-export * from "./topups";
-export * from "./withdrawals";
+export * from "./repository/withdrawal.repository";
+export * from "./services/commands/request-stripe-topup.service";
+export * from "./services/commands/request-withdrawal.service";
+export * from "./services/providers/stripe-topup.service";
+export * from "./services/providers/stripe-withdrawal.service";
+export * from "./services/shared/settle-topup.service";
+export * from "./services/shared/wallet-hold.service";
+export * from "./services/shared/wallet.service";
+export * from "./services/shared/withdrawal.service";
+export * from "./services/webhooks/stripe-connect.service";
+export * from "./services/webhooks/stripe-payout.service";
+export * from "./services/webhooks/stripe-topup-webhook.service";
+export * from "./services/webhooks/stripe-webhook.service";
+export * from "./services/workers/execute-withdrawal.service";
+export * from "./services/workers/reconcile-topup.service";
+export * from "./services/workers/sweep-withdrawals.service";
