+++ apps/server/src/domain/wallets/repository/wallet-hold.repository.ts
@@ -8,40 +8,7 @@ import { Prisma } from "@/infrastructure/prisma";
 import type { CreateWalletHoldInput, WalletHoldRow } from "../models";
 
 import { WalletHoldRepositoryError } from "../domain-errors";
-
-const selectWalletHoldRow = {
-  id: true,
-  walletId: true,
-  withdrawalId: true,
-  rentalId: true,
-  amount: true,
-  status: true,
-  reason: true,
-  releasedAt: true,
-  settledAt: true,
-  forfeitedAt: true,
-  createdAt: true,
-  updatedAt: true,
-} satisfies PrismaTypes.WalletHoldSelect;
-
-function toWalletHoldRow(
-  row: PrismaTypes.WalletHoldGetPayload<{ select: typeof selectWalletHoldRow }>,
-): WalletHoldRow {
-  return {
-    id: row.id,
-    walletId: row.walletId,
-    withdrawalId: row.withdrawalId,
-    rentalId: row.rentalId,
-    amount: row.amount,
-    status: row.status,
-    reason: row.reason,
-    releasedAt: row.releasedAt,
-    settledAt: row.settledAt,
-    forfeitedAt: row.forfeitedAt,
-    createdAt: row.createdAt,
-    updatedAt: row.updatedAt,
-  };
-}
+import { selectWalletHoldRow, toWalletHoldRow } from "./wallet-hold.mappers";
 
 export type WalletHoldRepo = {
   create: (
