--- apps/server/src/domain/wallets/topups/index.ts
@@ -1,8 +1,0 @@
-export * from "./domain-errors";
-export * from "./models";
-export * from "./repository/payment-attempt.repository";
-export * from "./services/reconcile-topup.service";
-export * from "./services/request-stripe-topup.service";
-export * from "./services/settle-topup.service";
-export * from "./services/stripe-topup-webhook.service";
-export * from "./services/stripe-topup.service";
