--- apps/server/src/domain/wallets/withdrawals/index.ts
@@ -1,10 +1,0 @@
-export * from "./domain-errors";
-export * from "./models";
-export * from "./repository/withdrawal.repository";
-export * from "./services/execute-withdrawal.service";
-export * from "./services/request-withdrawal.service";
-export * from "./services/stripe-connect.service";
-export * from "./services/stripe-payout.service";
-export * from "./services/stripe-withdrawal.service";
-export * from "./services/sweep-withdrawals.service";
-export * from "./services/withdrawal.service";
